rewrote single IMU simulation using new interfaces
